pyproject.toml
[tool.poetry]
name = "nonebot-plugin-splatoon3-nso"
version = "1.4.1"
description = "一个基于nonebot2框架的splatoon3游戏nso数据查询插件"
authors = ["cypas <ayano05@outlook.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_splatoon3_nso"}]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4="^4.12.2"
httpx="^0.24.1"
nonebot2="^2.2.0"
pydantic = ">=1.0.0,<3.0.0,!=2.5.0,!=2.5.1"
nonebot-adapter-onebot = "^2.4.1"
nonebot-adapter-telegram= "^0.1.0b16"
nonebot-adapter-kaiheila = "^0.3.1"
nonebot-adapter-qq = "^1.4.1"
nonebot-plugin-apscheduler = "^0.4.0"
nonebot_plugin_htmlrender = "^0.3.0"
SQLAlchemy = "^2.0.25"
playwright = "^1.40.0"
Pillow="9.5.0"


[tool.poetry.group.test.dependencies]
nonebug = "^0.3.5"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"